# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReplayBug API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "CLI secret project token (rb_sk_…). Send as Authorization: Bearer <token>. Never send credentials in query strings."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/cli/project": {
      "get": {
        "tags": ["CLI"],
        "description": "Resolve the owning project for a CLI secret token. Bearer-only authentication; never send credentials in query strings.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "projectId",
                    "projectName",
                    "projectSlug",
                    "workspaceId",
                    "workspaceName",
                    "timezone"
                  ],
                  "properties": {
                    "projectId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "projectName": {
                      "type": "string"
                    },
                    "projectSlug": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "workspaceName": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cli/releases": {
      "post": {
        "tags": ["CLI"],
        "description": "Create a project-scoped release. Idempotent: identical re-creates return 200 with created=false; differing identity metadata returns 409 RELEASE_VERSION_CONFLICT. Bearer-only authentication.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["version"],
                "properties": {
                  "version": {
                    "type": "string"
                  },
                  "commitSha": {
                    "type": "string"
                  },
                  "repositoryUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["release", "created"],
                  "properties": {
                    "release": {
                      "type": "object",
                      "required": [
                        "id",
                        "version",
                        "commit",
                        "repositoryUrl",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "version": {
                          "type": "string"
                        },
                        "commit": {
                          "type": ["null", "string"]
                        },
                        "repositoryUrl": {
                          "type": ["null", "string"]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "created": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["release", "created"],
                  "properties": {
                    "release": {
                      "type": "object",
                      "required": [
                        "id",
                        "version",
                        "commit",
                        "repositoryUrl",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "version": {
                          "type": "string"
                        },
                        "commit": {
                          "type": ["null", "string"]
                        },
                        "repositoryUrl": {
                          "type": ["null", "string"]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "created": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["CLI"],
        "description": "List project-scoped releases in deterministic order (created_at, then version) with artifact counts. Bearer-only authentication; never exposes credentials.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["releases"],
                  "properties": {
                    "releases": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "version",
                          "commit",
                          "artifactCount",
                          "createdAt"
                        ],
                        "properties": {
                          "version": {
                            "type": "string"
                          },
                          "commit": {
                            "type": ["null", "string"]
                          },
                          "artifactCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "requestId"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}